requirements.txt
pycryptodome
django
django-hcaptcha
bfa
websockets